Fix undo to undo commands in correct order
rename staffinstrumenteventhandler_hack to prepare for merge

Fix errors after merge, undoredo for pasting volumes

get that copy paste button made a little more nicely, and have it not appear in the arranger tab

fix up clipboard sprites. thanks to seymour schlong and SupraTheHedgehog
