+++ src/smp/ImageIndex.java
@@ -149,6 +149,7 @@ public enum ImageIndex {
     LOOP_PRESSED, LOOP_RELEASED, LOOP_LABEL,
     MUTE_PRESSED, MUTE_RELEASED, MUTE_LABEL,
     MUTE_A_PRESSED, MUTE_A_RELEASED, MUTE_A_LABEL,
+    CLIPBOARD_PRESSED, CLIPBOARD_RELEASED, CLIPBOARD_LABEL,
 
     SCROLLBAR_LEFT1, SCROLLBAR_LEFT2,
     SCROLLBAR_LEFT1_PRESSED, SCROLLBAR_LEFT2_PRESSED,
